refactoring: extract data gestion to component for Deezer pages
