+++ pages/deezer/artist.tsx
@@ -1,29 +1,11 @@
-import { Suspense } from "react";
-import useSWR from "swr";
 import Head from "next/head";
-import dynamic from "next/dynamic";
 
 import type { NextPage } from "next";
-import { UserTopArtist } from "types/deezer";
-import { ErrorProps } from "next/error";
 
-import { Layout, Loader } from "@components/index";
-import styles from "@styles/Pages/global.module.scss";
-
-const DeezerArtistCard = dynamic(async () => await import("@components/Deezer/DeezerArtistCard"), { suspense: true });
-
-const Error = dynamic(async () => await import("next/error"));
+import { Layout } from "@components/index";
+import { ArtistWrapper } from "@components/Deezer/ArtistWrapper";
 
 const Track: NextPage = () => {
-  const { container } = styles;
-
-  const fetcher = async (url: string): Promise<UserTopArtist> => {
-    const res = await fetch(url);
-    return await res.json();
-  };
-
-  const { data, error } = useSWR<UserTopArtist, ErrorProps>("/api/deezer/artists", fetcher);
-
   return (
     <>
       <Head>
@@ -33,18 +15,7 @@ const Track: NextPage = () => {
         <link rel="preload" href="/api/deezer/artists" as="fetch" crossOrigin="anonymous" />
       </Head>
       <Layout>
-        <section className={container}>
-          {error != null && <Error statusCode={error.statusCode} />}
-          {data !== undefined &&
-            data.data.length > 0 &&
-            data.data.map((item, index) => {
-              return (
-                <Suspense fallback={<Loader />} key={item.id}>
-                  <DeezerArtistCard index={index + 1} items={item} />
-                </Suspense>
-              );
-            })}
-        </section>
+        <ArtistWrapper />
       </Layout>
     </>
   );
